+++ apps/ledger-live-mobile/src/index.tsx
@@ -79,7 +79,7 @@ import {
 } from "@ledgerhq/live-common/exchange/swap/hooks/index";
 import useAccountsWithFundsListener from "@ledgerhq/live-common/hooks/useAccountsWithFundsListener";
 import { updateIdentify } from "./analytics";
-import { getFeature, useFeature } from "@ledgerhq/live-common/featureFlags/index";
+import { FeatureToggle, getFeature, useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { StorylyProvider } from "./components/StorylyStories/StorylyProvider";
 import { useSettings } from "~/hooks";
 import AppProviders from "./AppProviders";
@@ -91,6 +91,7 @@ import { walletSelector } from "~/reducers/wallet";
 import { exportWalletState, walletStateExportShouldDiffer } from "@ledgerhq/live-wallet/store";
 import { registerTransports } from "~/services/registerTransports";
 import { useDeviceManagementKitEnabled } from "@ledgerhq/live-dmk-mobile";
+import { StoragePerformanceOverlay } from "./newArch/storage/screens/PerformanceMonitor";
 
 if (Config.DISABLE_YELLOW_BOX) {
   LogBox.ignoreAllLogs();
@@ -249,6 +250,9 @@ function App() {
       <PerformanceConsole />
       <DebugTheme />
       <Modals />
+      <FeatureToggle featureId="llmMmkvMigration">
+        <StoragePerformanceOverlay />
+      </FeatureToggle>
     </GestureHandlerRootView>
   );
 }