+++ apps/web-tools/live-common-setup.ts
@@ -74,4 +74,5 @@ setSupportedCurrencies([
   "zenrock",
   "sui",
   "mina",
+  "babylon",
 ]);