+++ libs/ledger-live-common/src/e2e/swap.ts
@@ -1,20 +1,20 @@
-import { Swap } from "./models/Swap";
+import { Account } from "./enum/Account";
 import axios from "axios";
 
-export async function getMinimumSwapAmount(swap: Swap) {
+export async function getMinimumSwapAmount(AccountFrom: Account, AccountTo: Account) {
   try {
     const { data } = await axios({
       method: "GET",
       url: `https://swap-stg.ledger-test.com/v5/quote`,
       params: {
-        from: swap.accountToDebit.currency.id,
-        to: swap.accountToCredit.currency.id,
+        from: AccountFrom.currency.id,
+        to: AccountTo.currency.id,
         amountFrom: 0.0001,
-        addressFrom: swap.accountToDebit.address,
+        addressFrom: AccountFrom.address,
         fiatForCounterValue: "USD",
         slippage: 1,
         networkFees: 0.001,
-        networkFeesCurrency: swap.accountToDebit.currency.speculosApp.name.toLowerCase(),
+        networkFeesCurrency: AccountTo.currency.speculosApp.name.toLowerCase(),
         displayLanguage: "en",
         theme: "light",
         "providers-whitelist": "changelly,exodus,thorswap,uniswap,cic",