+++ libs/ledger-live-common/src/featureFlags/defaultFeatures.ts
@@ -91,6 +91,7 @@ export const CURRENCY_DEFAULT_FEATURES = {
   currencySonic: DEFAULT_FEATURE,
   currencySui: DEFAULT_FEATURE,
   currencyMina: DEFAULT_FEATURE,
+  currencyBabylon: DEFAULT_FEATURE,
 };
 
 /**
@@ -599,6 +600,7 @@ export const DEFAULT_FEATURES: Features = {
       send_flow: true,
     },
   },
+  llNftSupport: DEFAULT_FEATURE,
 };
 
 // Firebase SDK treat JSON values as strings