+++ libs/ledgerjs/packages/hw-app-celo/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-polkadot
 
+## 6.33.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.3
+
+## 6.33.1-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-app-eth@6.45.3-next.0
+
 ## 6.33.0
 
 ### Minor Changes